Move spec to umbrella file
